Directly execute scripts by default, only use a shell if set to
Previously, the script exporter ran all configured scripts as 'SHELL
script [args ...]'. This makes the shell (/bin/sh by default) always
attempt to interpret the script as a shell script for itself, even if
the script is really an executable program or a shell script for
another shell (for example, on a system where /bin/sh is not Bash and
the script starts with '#! /bin/bash'). If this happens, various
things go wrong, sometimes very badly.

We now support directly executing configured scripts and default to
this. Directly executing scripts works for compiled programs and any
script that starts with '#!', which is the usual case. This is
controlled by whether there is a '-config.shell' command line argument
set; by default there is not.

(While defaulting to directly executing scripts is a potentially
breaking change for any current script that doesn't start with '#!'
or isn't executable, it is Chris Siebenmann's opinion that this is the
correct default. The current behaviour is surprising in several
respects.)