Provide some internal metrics for script_exporter
We now provide the standard set of Prometheus Go client metrics and
some additional script exporter specific ones on /metrics, with the
previous script endpoint moved to /probes. For additional metrics, we
provide per-script metrics on total requests, in-flight requests, and
duration of requests, and total count and total duration broken down
by HTTP status code and method. Requests to /metrics are not
authenticated, but requests to /probes continue to be authenticated if
authentication has been configured.

In a breaking change, the -web.telemetry-path option has been
removed (and the script endpoint changed to /probes). The script
endpoint and metrics endpoint are both hard-coded, matching the
behaviour of the Blackbox exporter.

This adds a dependency on github.com/prometheus/client_golang.

Closes #8