Expose Prometheus scrape timeout in the script's environment
When Prometheus scrapes an endpoint, it exposes its timeout for the
scrape in the request's HTTP headers. We now extract this information
from the incoming request and pass it through to scripts that we run
in two environment variables, $SCRIPT_TIMEOUT and $SCRIPT_DEADLINE, so
that they can use it to set their own timeouts instead of having to
guess. We copy Blackbox's behaviour of allowing the maximum timeout to
be set in the configuration file, either globally or on a per-script
basis. We also copy Blackbox in having a timeout offset that is
subtracted from the Prometheus-supplied timeout to allow for various
bits of overhead. For testing, we also allow the timeout to be
specified directly in the URL as a query parameter, `timeout`. This
currently overrides the HTTP header if it contains a valid value.

An example script has been added that simply shows the values of the
timeout environment variables (and whether or not they're present),
and the example configuration and README.md updated to discuss all of
this.