Optionally enforce scrape timeouts by killing scripts when necessary
We can now attempt to enforce scrape timeouts from Prometheus (or the
`timeout` URL query parameter) by killing the main process for a
script when the timeout deadline is reached. This is implemented by
using exec.CommandContext with a context with a deadline. Whether or
not timeout enforcement is on for a given script is exposed to it in
the new environment variable $SCRIPT_TIMEOUT_ENFORCED, so a script can
take appropriate precautions.

The intention (or goal) is that if you are just going to wrap scripts
you run with 'timeout $SCRIPT_TIMEOUT' anyway, script_exporter might
as well do the work for you.

The signal that scripts are terminated with is currently not settable,
and we make no attempt to do clever things like put every script in
its own process group and then kill the entire process group so that
we sweep up sub-processes as well. (This is not supported by
exec.CommandContext, so we would have to implement our own version.)