Remove the global 'timeouts' section from the configuration file
This implements the change requested in feedback in issue #11, where
now we only have a 'timeout' section in each script with no global
'timeouts' section. The actual implementation of this is a minimal
code change that leaves pkg/config/config.go with a separate 'timeout'
struct type instead of having an anonymous struct in the Scripts
anonymous struct.