:warning: Fix Environment Variable Handling
In the readme we had an example on how to use the environment variables
via key value pairs, but in the code we had it implemented as slice.

To fix this and because I prefer the key value pairs the code was
adjusted to use a "map[string]string" instead of a slice of strings to
specify the environment variables.