Add Cache Option
It is now possible to cache the output of an script via the newly added
"cacheDuration" option. If this option is provided the exporter will
return the output from the cache instead of executing the script again
as long as the output is not stale.

If no cache duration is provided or if the provided cache duration can
not be parsed, the output of a script will not be cached.