Add Flag to Enable Logging of Environment Variable
Until now we always logged environment variables used by the exporter to
run a script. This means that it was possible that the logs also
contained sensible information from these environment variables
(see #91). This is now changed an we do not log environment variables
anymore. For debugging this behaviour can be overwritten by setting the
new "-log.env" flag to "true".