Define Additional Scripts via Separate Config Files
It is now possible to define additional scripts for the exporter via
separate config files. For that a new option `scripts_configs` was
added, which can be used to set a list of files, which are containing
additional script configuration.

The configuration must look as follows:

```
script_configs:
  - /path/to/file.yaml
```

The content of the additional scripts configuration file, can only
contain a list of scripts, e.g.

```yaml
- name: test
  command: ./examples/test.sh
```