Fix Race Condition when Cache is Used
When to requests were accessing the cache during the same time, there
was a race condition because of the concurrent access to the map which
holds the cached entries.

This is now fixed by adding a lock for the cache, so that multiple
requests can not access the cache in parallel.