Use GitHub Container Registry to Store Helm Chart
Instead of building and uploading the Helm Chart for the script_exporter
to the https://github.com/ricoberger/helm-charts repository, we are now
using the GitHub Container Registry to store the Helm chart.

Starting with the new version 1.0.0 the Helm chart will only be
available at `ghcr.io/ricoberger/charts/script-exporter`.