Helm Chart: Adjust ServiceMonitors
This commit adjusts the creation of the ServiceMonitors in the Helm
chart and is a breaking change compared to the last Helm chart version.

The `serviceMonitor.selfMonitor` values are moved to
`selfServiceMonitor` and the available values are adjusted to follow a
similar naming as it is used for the `serviceMonitor` values.

The `honorLabels` option is now used within the `selfServiceMonitor` and
`serviceMonitor`.