Don't Wait for IO Pipes to be Closed
When a executed script spawned a child process (e.g. `sleep`) the
timeout was not enforced, because `cmd.Wait()` would wait for the script
to exit and for all IO pipes to be closed. This is now fixed by adding a
`cmd.WaitDelay` of 1ms to kill the process even, if the IO pipes are not
closed.