Fix Build for Binaries
In the continuous delivery GitHub Action for the binaries the `GOOS`,
`GOARCH` and `CGO_ENABLED` environment variables were not set, so that
the binaries were not build correctly.

Additionally we had to add a build constraint to the `cmd/main_unix.go`
file, so that the build is working correctly for Windows.