Add Support to Load Configuration from URL
It is now possible to specify a url in the `--config.files` flag, so
that the configuration is loaded from the url instead of a local file.
To be able to reload the configuration regularly from the url, a new
flag `--config.reload-interval` has been added. The default value is 1h,
which means that the configuration is reloaded every hour.